Revert "refactor: Added results uploading"
This reverts commit e4cf8d15563dffacc9ecd8370b413db91e7e8634.
