+++ src/main.ts
@@ -1,7 +1,7 @@
 import { DeviceBase } from "./device";
 import { getDevices } from "./devices";
-import { downloadBuilds, versionBundle, uploadResults } from "./cloudAccess";
-const core = require('@actions/core');
+import { downloadBuilds, versionBundle } from "./downloadBuilds";
+import * as core from '@actions/core';
 import * as path from 'path';
 
 export const packageName = core.getInput('packageName');
@@ -28,9 +28,9 @@ async function main() {
   //Step 4: Wait for execution to be completed on all the connected devices
   waitForAllDevicesToComplete(devices).then(() => {
     allDevicesCompleted(devices);
-    const resultsName = `Results_${versionBundle}`;
+    const resultsName = `Results_${versionBundle}`
     const targetPath = path.join(__dirname, '../', `${resultsName}`);
-    uploadResults(targetPath);
+    core.setOutput('resultsPath', targetPath);
     console.log("All devices have completed their execution.");
   }).catch((err) => {
     console.error("An error occurred while waiting for devices to complete:", err);
